+++ mint_background_switcher/__init__.py
@@ -1,5 +1,5 @@
 """Mint Background Switcher package."""
 
-__version__ = "0.1.16"
+__version__ = "0.1.17"
 APP_ID = "mint-background-switcher"
 APP_NAME = "Mint Background Switcher"